# docs/swagger.yaml
basePath: /
definitions:
...
  internal_rest_handlers_product.Product:
    properties:
      description:
        type: string
      image_url:
        type: string
      name:
        type: string
      price:
        minimum: 0
        type: integer
      quantity:
        minimum: 0
        type: integer
    required:
    - name
    - price
    - quantity
    type: object
...
  user.Profile:
    properties:
      avatar_url:
        example: /profile/1/photo
        type: string
      id:
        example: 1
        type: integer
      phone_number:
        example: "+998901234567"
        type: string
      pinfl:
        example: "12345678901234"
        type: string
      role:
        example: user
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: REST API для управления заказами, продуктами и пользователями.
  license:
    name: MIT
  termsOfService: http://example.com/terms/
  title: Order Service API
  version: "1.0"
paths:
  /api/v1/admin/users:
    get:
      description: Админский доступ. Возвращает список всех зарегистрированных пользователей
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/user.Profile'
            type: array
        "403":
          description: forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: internal error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получение списка всех пользователей
      tags:
      - Admin - Users
...
  /api/v1/me:
    get:
      description: Возвращает информацию о текущем пользователе по JWT-токену
      produces:
      - application/json
      responses:
        "200":
          description: Профиль пользователя
          schema:
            $ref: '#/definitions/user.Profile'
        "401":
          description: unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: user not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: internal error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получение профиля текущего пользователя
      tags:
      - User
    patch:
      consumes:
      - multipart/form-data
      description: Обновляет PINFL и аватар текущего пользователя
      parameters:
      - description: ПИНФЛ пользователя
        in: formData
        name: pinfl
        type: string
      - description: Аватар (изображение)
        in: formData
        name: avatar
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: profile updated
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: user not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: internal error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Обновление профиля пользователя
      tags:
      - User
...
  /api/v1/products:
    get:
      description: Возвращает список всех доступных продуктов
      responses:
        "200":
          description: 'products: []Product'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all products
      tags:
      - products
    post:
      description: Добавляет новый продукт в каталог
      parameters:
      - description: Product object
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/internal_rest_handlers_product.Product'
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Add new product (admin)
      tags:
      - products
  /api/v1/products/{id}:
    delete:
      description: Удаляет продукт по ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete product by ID (admin)
      tags:
      - products
    get:
      description: Возвращает продукт по его ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: 'product: Product'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get product by ID
      tags:
      - products
    put:
      description: Обновляет информацию о продукте по ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated product
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/internal_rest_handlers_product.Product'
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update product by ID (admin)
      tags:
      - products
  /profile/{id}/photo:
    get:
      description: Возвращает файл изображения профиля по ID пользователя
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      produces:
      - image/jpeg
      - image/png
      - image/webp
      responses:
        "200":
          description: Изображение аватара
          schema:
            type: file
        "404":
          description: photo not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получение аватара пользователя
      tags:
      - User
swagger: "2.0"